+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 Select a Spreadsheet Light version below to view the changelog history:
 
-- [Spreadsheet Light 0.1.6](#V0.1.6) **Current**
+- [Spreadsheet Light 0.1.6](#V0.1.7) **Current**
+- [Spreadsheet Light 0.1.6](#V0.1.6)
+
+## V1.7
+
+- Fixing issue when an empty value is followed by the end of line
 
 ## V1.6